requirements.txt
Django==2.2.3
pytz==2019.1
sqlparse==0.3.0
django-filter==2.3.0
djangorestframework==3.11.1
